More fixes and improving selection deep linkage
