+++ package.json
@@ -22,6 +22,7 @@
     "node-sass": "^6.0.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-icons": "^4.3.1",
     "react-lorem-ipsum": "^1.4.9",
     "react-rb-auth": "^3.1.8",
     "react-redux": "^7.2.5",
