+++ src/components/radar/Filter.tsx
@@ -1,108 +1,24 @@
-import React, { ChangeEventHandler, useEffect, useState } from 'react';
-import { Connect } from 'redux-auto-actions';
+import React, { useState } from 'react';
+// import { FaCog } from 'react-icons/fa';
+import { FiSettings } from 'react-icons/fi';
 
-import { RadarUtilities } from '../../radar/utilities/Utilities';
-import { actions, selectors } from '../../store/radar/radar.actions';
-import { GlobalState } from '../../store/state';
+import { RawFilter } from './RawFilter';
+import styles from './Filter.module.scss';
 
-export const Filter = Connect<GlobalState, unknown>()
-  .stateAndDispatch(
-    (state) => ({
-      blips: selectors(state).blips,
-      disasterTypeFilter: selectors(state).disasterTypeFilter,
-      useCaseFilter: selectors(state).useCaseFilter,
-    }),
-    {
-      setUseCaseFilter: actions.setUseCaseFilter,
-      setDisasterTypeFilter: actions.setDisasterTypeFilter,
-    }
-  )
-  .withComp(({ blips, disasterTypeFilter, useCaseFilter, setUseCaseFilter, setDisasterTypeFilter }) => {
-    const [disasterTypes, setDisasterTypes] = useState<SelectableItem[]>([]);
-    const [useCases, setUseCases] = useState<SelectableItem[]>([]);
+export const Filter: React.FC = () => {
+  const [collapsed, setCollapsed] = useState(true);
 
-    useEffect(() => {
-      if (blips && blips?.length > 0) {
-        const newUseCases = RadarUtilities.getUseCases(blips);
-        setUseCases(newUseCases);
+  const onFilterToggle = () => setCollapsed(!collapsed);
 
-        const newDisasterTyes = RadarUtilities.getDisasterTypes(blips);
-        setDisasterTypes(newDisasterTyes);
-      }
-    }, [blips]);
-
-    const [selectedDisasterType, setSelectedDisasterType] = useState<string>(
-      disasterTypeFilter === null ? 'all' : disasterTypeFilter
-    );
-    const [selectedUserCase, setSelectedUserCase] = useState<string>(useCaseFilter === null ? 'all' : useCaseFilter);
-    const onDisasterTypeChange: ChangeEventHandler<HTMLSelectElement> = (e) => setSelectedDisasterType(e.target.value);
-    const onUseCaseChange: ChangeEventHandler<HTMLSelectElement> = (e) => setSelectedUserCase(e.target.value);
-
-    const onFilterHnalder = () => {
-      // selected?
-      setUseCaseFilter(selectedUserCase);
-      setDisasterTypeFilter(selectedDisasterType);
-    };
-
-    return (
-      <div
-        style={{
-          margin: 10,
-          marginTop: 30,
-          borderStyle: 'solid',
-          borderColor: 'black',
-          borderWidth: 1,
-          padding: 20,
-          paddingTop: 20,
-          borderRadius: 5,
-          maxWidth: 300,
-        }}
-      >
-        <div>Customize Radar</div>
-
-        <div style={{ paddingTop: 20 }}>
-          <select id="Select1" style={{ width: '100%' }} onChange={onDisasterTypeChange} value={selectedDisasterType}>
-            <option value="all">Show all countries</option>
-            {disasterTypes.map((item) => (
-              <option key={item.uuid} value={item.name}>
-                {item.name}
-              </option>
-            ))}
-          </select>
-        </div>
-
-        <div style={{ paddingTop: 20 }}>
-          <select id="Select2" style={{ width: '100%' }} onChange={onUseCaseChange} value={selectedUserCase}>
-            <option value="all">Show all Use Cases</option>
-            {useCases.map((item) => (
-              <option key={item.uuid} value={item.name}>
-                {item.name}
-              </option>
-            ))}
-          </select>
-        </div>
-
-        <small style={{ padding: 10, fontSize: 10, textAlign: 'left', float: 'left', width: '100%' }}>
-          todo: create filter generalization
-        </small>
-
-        <div style={{ paddingTop: 20 }}>
-          <button
-            type="button"
-            style={{
-              borderColor: 'black',
-              borderWidth: 1,
-              borderStyle: 'solid',
-              padding: '10px 20px',
-              backgroundColor: 'whitesmoke',
-              cursor: 'pointer',
-              borderRadius: 5,
-            }}
-            onClick={onFilterHnalder}
-          >
-            Filter
-          </button>
-        </div>
+  return (
+    <div className={styles.wrapper}>
+      <div className={styles.button} onClick={onFilterToggle} onKeyUp={onFilterToggle} role="button" tabIndex={0}>
+        <FiSettings size={30} color="white" />
+        <div style={{ paddingLeft: 10, color: 'white' }}>Customize radar</div>
       </div>
-    );
-  });
+      <div className={collapsed ? styles.filterBoxCollapsed : styles.filterBoxUncollapsed}>
+        <RawFilter />
+      </div>
+    </div>
+  );
+};
